Redo code execution, in particular return messages
Code is executed phrase by phrase and return messages built up into
a list.  These are then individually formatted into html and
sent to ipython.

For html formatting compiler outputs are in gray and errors in
red.  Each output is limited to a css max-height, clicking on the
message will expand/collapse it.  This avoids the compiler messages
dominating the notebook output, especially for modules.