Fix notebook start by custom msg handling
Problem
-------
Kernel doesn't handle "comm_open" [custom message type][1], which is a
part of a new mechanism in Jupyter 4.1, a mechanism of custom messgaes.

How to reproduce? When notebook is opened kernel crashes as "comm_open"
is sent. But if manually restart the kernel from the notebook menu this
message is not sent so we can get a notebook in a working state.

Solution
--------
Solution doesn't support custom messages but only provides the stubs to
not crash kernel.

Tested
------
Manually. Notebook starts with successfully initialyzed kernel.

[1]: http://jupyter-client.readthedocs.org/en/latest/messaging.html#opening-a-comm